+++ src/main/java/org/example/service/UserTraining.java
@@ -70,7 +70,7 @@ public class UserTraining {
      * @param username имя текущего пользователя
      * @param session текущая сессия
      */
-    public void saveUsersTrainingTime(double time, String username, Session session) {
+    public void saveUsersTrainingTime(int time, String username, Session session) {
         try {
             if (username != null) {
                 UserEntity user = userDao.getUserByUsername(username, session);
@@ -91,7 +91,7 @@ public class UserTraining {
      * @param session текущая сессия
      * @return время тренировки
      */
-    public double getUserTrainingTime(String username, Session session) {
+    public int getUserTrainingTime(String username, Session session) {
         try {
             if (username != null) {
                 UserEntity user = userDao.getUserByUsername(username, session);
@@ -101,6 +101,6 @@ public class UserTraining {
             logger.error(e.getMessage(), e);
             inputOutput.output(e.getMessage());
         }
-        return 0.0;
+        return 0;
     }
 }
